Add audio options in Pause screen
