+++ src/ui/PauseOverlay.java
@@ -4,48 +4,31 @@ import gameStates.Playing;
 import gameStates.Gamestate;
 import main.Game;
 import utilz.LoadSave;
-import utilz.Constant;
-import utilz.LoadSave;
 
 import java.awt.*;
 import java.awt.event.MouseEvent;
 import java.awt.image.BufferedImage;
 
-import static utilz.Constant.UI.PauseButtons.*;
-import static utilz.Constant.UI.URMButtons.*;
-import static utilz.Constant.UI.VolumeButtons.*;
+import static utilz.Constant.UI.URMButtons.*;;
 
 
 public class PauseOverlay {
     private Playing playing;
     private BufferedImage backgroundImg;
     private int bgX, bgY, bgW, bgH;
-    private SoundButton musicButton, sfxButton;
+    private AudioOptions audioOptions;
     private UrmButton menuB, replayB, unpauseB;
-    private VolumeButton volumeButton;
 
 
     public PauseOverlay(Playing playing) {
         this.playing = playing;
         loadBackground();
-        createSoundButtons();
+        audioOptions = playing.getGame().getAudioOptions();
+
         createUrmButtons();
-        createVolumeButton();
-    }
-    private void createVolumeButton() {
-        int vX = (int) (309 * Game.SCALE);
-        int vY = (int) (278 * Game.SCALE);
-        volumeButton = new VolumeButton(vX, vY, SLIDER_WIDTH, VOLUME_HEIGHT);
-    }
-
-    private void createSoundButtons() {
-        int soundX = (int) (450 * Game.SCALE);
-        int musicY =  (int) (140 * Game.SCALE);
-        int sfxY = (int) (186 * Game.SCALE);
-        musicButton = new SoundButton(soundX, musicY, SOUND_SIZE, SOUND_SIZE);
-        sfxButton = new SoundButton(soundX, sfxY, SOUND_SIZE, SOUND_SIZE);
 
     }
+
     private void createUrmButtons() {
         int menuX = (int) (313 * Game.SCALE);
         int replayX = (int) (387 * Game.SCALE);
@@ -67,13 +50,11 @@ public class PauseOverlay {
     }
 
     public void update(){
-        musicButton.update();
-        sfxButton.update();
         menuB.update();
         replayB.update();
         unpauseB.update();
 
-        volumeButton.update();
+        audioOptions.update();
     }
 
 
@@ -81,48 +62,30 @@ public class PauseOverlay {
         //Background
         g.drawImage(backgroundImg, bgX, bgY, bgW, bgH, null);
 
-        //Sound buttons
-        musicButton.draw(g);
-        sfxButton.draw(g);
         // UrmButtons
         menuB.draw(g);
         replayB.draw(g);
         unpauseB.draw(g);
 
-        // Volume Button
-        volumeButton.draw(g);
+        // Audio Options
+        audioOptions.draw(g);
     }
 
 
-
     public void mousePressed(MouseEvent e) {
-        if(isIn(e, musicButton)){
-            musicButton.setMousePressed(true);
-        } else if (isIn(e, sfxButton)){
-            sfxButton.setMousePressed(true);
-        }
-        else if (isIn(e, menuB))
+        if (isIn(e, menuB))
             menuB.setMousePressed(true);
         else if (isIn(e, replayB))
             replayB.setMousePressed(true);
         else if (isIn(e, unpauseB))
             unpauseB.setMousePressed(true);
-        else if (isIn(e, volumeButton))
-            volumeButton.setMousePressed(true);
+        else
+            audioOptions.mousePressed(e);
     }
 
 
-
-
     public void mouseReleased(MouseEvent e) {
-        if(isIn(e, musicButton)){
-            if(musicButton.isMousePressed())
-                musicButton.setMuted(!musicButton.isMuted()); //set it to the opposite state of muted
-        } else if (isIn(e, sfxButton)){
-            if(sfxButton.isMousePressed())
-                sfxButton.setMuted(!sfxButton.isMuted());
-        }
-        else if (isIn(e, menuB)) {
+        if (isIn(e, menuB)) {
             if (menuB.isMousePressed()) {
                 Gamestate.state = Gamestate.MENU;
                 playing.unpauseGame();
@@ -135,47 +98,33 @@ public class PauseOverlay {
         } else if (isIn(e, unpauseB)) {
             if (unpauseB.isMousePressed())
                 playing.unpauseGame();
-        }
+        } else
+            audioOptions.mouseReleased(e);
 
-        musicButton.resetBools();
-        sfxButton.resetBools();
         menuB.resetBools();
         replayB.resetBools();
         unpauseB.resetBools();
-        volumeButton.resetBools();
-
     }
 
 
     public void mouseDragged(MouseEvent e) {
-        if (volumeButton.isMousePressed()) {
-            volumeButton.changeX(e.getX());
-        }
-
+        audioOptions.mouseDragged(e);
     }
 
 
-
     public void mouseMoved(MouseEvent e) {
-        musicButton.setMouseOver(false);
-        sfxButton.setMouseOver(false);
+        menuB.setMouseOver(false);
         replayB.setMouseOver(false);
         unpauseB.setMouseOver(false);
-        volumeButton.setMouseOver(false);
 
-        if(isIn(e, musicButton)){
-            musicButton.setMouseOver(true);
-        } else if (isIn(e, sfxButton)){
-            sfxButton.setMouseOver(true);
-        }
-        else if (isIn(e, menuB))
+        if (isIn(e, menuB))
             menuB.setMouseOver(true);
         else if (isIn(e, replayB))
             replayB.setMouseOver(true);
         else if (isIn(e, unpauseB))
             unpauseB.setMouseOver(true);
-        else if (isIn(e, volumeButton))
-            volumeButton.setMouseOver(true);
+        else
+            audioOptions.mouseMoved(e);
 
     }
